+++ CHANGELOG.md
@@ -37,6 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/1061>)
 - Enhance the error message for datum stats to be more user friendly
   (<https://github.com/openvinotoolkit/datumaro/pull/1069>)
+- Refactor dataset.py to seperate DatasetStorage
+  (<https://github.com/openvinotoolkit/datumaro/pull/1073>)
 
 ### Bug fixes
 - Fix warnings in test_visualizer.py